pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inflex_sganfin"
version = "0.0.1"
authors = [
    { name="Álvaro García Tenorio", email="agarcia@economia.gob.es"},
]
description = "Flexible report generator"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
    'python-docx >= 1.1.2',
    'PyYAML >= 6.0.2',
]

[project.urls]
Homepage = "https://github.com/sgtyfianalisisfinanciero/inflex-sganfin"
